+++ app/src/test/java/app/knotwork/android/domain/usecases/ParkedRunResumerTest.kt
@@ -14,6 +14,7 @@ import app.knotwork.android.domain.repositories.PipelineRunRepository
 import app.knotwork.android.domain.repositories.SettingsRepository
 import app.knotwork.android.domain.services.ApprovalNotifier
 import app.knotwork.android.domain.services.ClarificationNotifier
+import app.knotwork.android.domain.services.RunOutcomeAnnouncer
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
@@ -42,6 +43,7 @@ class ParkedRunResumerTest {
     private lateinit var clarificationNotifier: ClarificationNotifier
     private lateinit var resumePipelineRunUseCase: ResumePipelineRunUseCase
     private lateinit var recordTriggerHitlEvent: RecordTriggerHitlEventUseCase
+    private lateinit var runOutcomeAnnouncer: RunOutcomeAnnouncer
     private lateinit var resumer: ParkedRunResumer
 
     @Before
@@ -53,6 +55,7 @@ class ParkedRunResumerTest {
         clarificationNotifier = mockk(relaxed = true)
         resumePipelineRunUseCase = mockk()
         recordTriggerHitlEvent = mockk(relaxed = true)
+        runOutcomeAnnouncer = mockk(relaxed = true)
         resumer = ParkedRunResumer(
             pendingInteractionRepository = pendingInteractionRepository,
             pipelineRunRepository = pipelineRunRepository,
@@ -61,6 +64,7 @@ class ParkedRunResumerTest {
             clarificationNotifier = clarificationNotifier,
             resumePipelineRunUseCase = resumePipelineRunUseCase,
             recordTriggerHitlEvent = recordTriggerHitlEvent,
+            runOutcomeAnnouncer = runOutcomeAnnouncer,
         )
         coEvery { settingsRepository.backgroundApprovalWindowHours } returns flowOf(24)
         coEvery { resumePipelineRunUseCase("run-1") } returns ResumeOutcome.Resumed